fix: Revert wECO mainnet to a dummy address
There are utility functions that iterate over the array of
TOKEN_SYMBOLS_MAP values and return the first symbol matching an address
for a particular chain ID. If chainIds are reused in this way, the
ordering of TOKEN_SYMBOLS_MAP determines which symbol would be returned.
This is generally undesirable, so avoid reusing token addresses in this
way.