feat: add native USDC + distinction between bridged USDC (#28)
I am not 100% sure if this is the best approach. But on the FE/API, I
now need to distinguish between native and bridged USDC. So far we have
only had the entry `USDC` for both bridged and native ones. Eventually,
it would make sense to replace the old `USDC` entry with the values of
`_USDC`. But as this would probably break other parts of the stack, I
opted to use the `_USDC` entry as a temporary workaround.

Alternatively, I could locally override the `TOKEN_SYMBOLS_MAP` instead
of publishing this change. Thoughts?

Addresses are taken from here
https://developers.circle.com/stablecoins/docs/usdc-on-main-networks

---------

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>