chore: Add Aleph Zero definitions (#75)
I'm not 100% sure how much we're going to need the AZERO/WAZERO
definitions, but have added them as a precaution. Non-ETH gas tokens
definitely expose some quirks in the way we resolve tokens in the FE and
the bots; that's tech debt that we'll need to reconcile before too long.